backend/requirements.txt
dnspython==2.2.1
fastapi==0.88.0
gunicorn==20.1.0
httptools==0.5.0
python-dotenv==0.21.0
requests==2.28.1
smart-open==6.3.0
smtplibaio==2.1.2.post1
sniffio==1.3.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
websockets==10.4
firebase_admin==4.5.0
pyrebase
itsdangerous
google-auth